requirements.txt
fastapi==0.110.1
uvicorn==0.29.0

grpcio==1.62.1
grpcio-tools==1.62.1

protobuf==4.25.3

pika==1.3.2
requests==2.31.0